Added examples of patterns
Added some example patterns on lists, tuples, and basic user-defined types.

I think the 'as-patterns' definitely belong since they sometimes throws
off beginners.

I also changed the description of the case expression to illustrate that the
alternatives are distinguished by 'patterns' and not 'expressions'. (You can't
match on (5 + 5) for example.

This is just a braindump and will likely need to be trimmed down. But since
there's currently space it seemed worthwhile.